Inline property and state
